Edit List button works, Cancel Edit button works, just can't click edit button twice in a row
